Implement loading, error capturing display, recommendation and evaluation logic

+++ backend/src/main/java/ai/preferred/cornac/repository/ExperimentRepository.java
@@ -8,7 +8,7 @@ import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface ExperimentRepository extends CrudRepository<Experiment, Long> {
+public interface ExperimentRepository extends CrudRepository<Experiment, Integer> {
     List<Experiment> findAll();
 
     Experiment findFirstByEndDateTimeIsNull();
